chore(eslint-plugin): extract FunctionSignature and enum comparison utils
